Refactor MerkleTreeGadget implementation
