+++ primitives/src/circuit/merkle_tree/mod.rs
@@ -6,17 +6,23 @@
 
 //! Trait definitions for a Merkle tree gadget.
 
-use crate::{
-    merkle_tree::{MerkleTreeScheme, UniversalMerkleTreeScheme},
-    rescue::RescueParameter,
-};
 use ark_ff::PrimeField;
 use jf_relation::{errors::CircuitError, BoolVar, Circuit, PlonkCircuit, Variable};
 
 mod rescue_merkle_tree;
 mod sparse_merkle_tree;
-use ark_std::vec::Vec;
+use ark_std::{string::ToString, vec::Vec};
+use typenum::{Unsigned, U3};
 
+use crate::{
+    merkle_tree::{
+        internal::{MerkleNode, MerklePath, MerkleProof},
+        prelude::RescueMerkleTree,
+        Element, Index, MerkleTreeScheme, NodeValue, ToTraversalPath, UniversalMerkleTreeScheme,
+    },
+    rescue::RescueParameter,
+};
+type NodeVal<F> = <RescueMerkleTree<F> as MerkleTreeScheme>::NodeValue;
 use super::rescue::RescueNativeGadget;
 
 /// Gadget for a Merkle tree
@@ -279,5 +285,157 @@ impl<F: RescueParameter> DigestAlgorithmGadget<F> for RescueDigestGadget {
     ) -> Result<Variable, CircuitError> {
         let zero = circuit.zero();
         Ok(RescueNativeGadget::<F>::rescue_sponge_no_padding(circuit, &[zero, pos, elem], 1)?[0])
+    }
+}
+
+/// Proof of membership
+pub trait MembershipProof<E, I, T>
+where
+    E: Element,
+    I: Index,
+    T: NodeValue,
+{
+    /// Get the tree height
+    fn tree_height(&self) -> usize;
+    /// Get index of element
+    fn index(&self) -> &I;
+    /// Get the element
+    fn elem(&self) -> Option<&E>;
+    /// Get the merkle path to the element
+    fn merkle_path(&self) -> &MerklePath<E, I, T>;
+}
+
+impl<E, I, T, Arity> MembershipProof<E, I, T> for MerkleProof<E, I, T, Arity>
+where
+    E: Element,
+    I: Index,
+    T: NodeValue,
+    Arity: Unsigned,
+{
+    fn tree_height(&self) -> usize {
+        self.tree_height()
+    }
+    fn index(&self) -> &I {
+        self.index()
+    }
+    fn elem(&self) -> Option<&E> {
+        self.elem()
+    }
+    fn merkle_path(&self) -> &MerklePath<E, I, T> {
+        &self.proof
+    }
+}
+
+impl<T> MerkleTreeGadget<T> for PlonkCircuit<T::NodeValue>
+where
+    T: MerkleTreeScheme,
+    T::MembershipProof: MembershipProof<T::NodeValue, T::Index, T::NodeValue>,
+    T::NodeValue: PrimeField + RescueParameter,
+    T::Index: ToTraversalPath<U3>,
+{
+    type MembershipProofVar = Merkle3AryMembershipProofVar;
+
+    type DigestGadget = RescueDigestGadget;
+
+    fn create_membership_proof_variable(
+        &mut self,
+        merkle_proof: &<T as MerkleTreeScheme>::MembershipProof,
+    ) -> Result<Merkle3AryMembershipProofVar, CircuitError> {
+        let path = merkle_proof
+            .index()
+            .to_traversal_path(merkle_proof.tree_height() - 1);
+
+        let elem = match merkle_proof.elem() {
+            Some(elem) => elem,
+            None => {
+                return Err(CircuitError::InternalError(
+                    "The proof doesn't contain a leaf element".to_string(),
+                ))
+            },
+        };
+
+        let elem_var = self.create_variable(*elem)?;
+
+        let nodes = path
+            .iter()
+            .zip(merkle_proof.merkle_path().iter().skip(1))
+            .filter_map(|(branch, node)| match node {
+                MerkleNode::Branch { value: _, children } => Some((children, branch)),
+                _ => None,
+            })
+            .map(|(children, branch)| {
+                let sib_branch1 = if branch == &0 { 1 } else { 0 };
+                let sib_branch2 = if branch == &2 { 1 } else { 2 };
+                Ok(Merkle3AryNodeVar {
+                    sibling1: self.create_variable(children[sib_branch1].value())?,
+                    sibling2: self.create_variable(children[sib_branch2].value())?,
+                    is_left_child: self.create_boolean_variable(branch == &0)?,
+                    is_right_child: self.create_boolean_variable(branch == &2)?,
+                })
+            })
+            .collect::<Result<Vec<Merkle3AryNodeVar>, CircuitError>>()?;
+
+        // `is_left_child`, `is_right_child` and `is_left_child+is_right_child` are
+        // boolean
+        for node in nodes.iter() {
+            // Boolean constrain `is_left_child + is_right_child` because a node
+            // can either be the left or the right child of its parent
+            let left_plus_right =
+                self.add(node.is_left_child.into(), node.is_right_child.into())?;
+            self.enforce_bool(left_plus_right)?;
+        }
+
+        Ok(Merkle3AryMembershipProofVar {
+            node_vars: nodes,
+            elem_var,
+        })
+    }
+
+    fn create_root_variable(
+        &mut self,
+        root: NodeVal<T::NodeValue>,
+    ) -> Result<Variable, CircuitError> {
+        self.create_variable(root)
+    }
+
+    fn is_member(
+        &mut self,
+        elem_idx_var: Variable,
+        proof_var: Merkle3AryMembershipProofVar,
+        root_var: Variable,
+    ) -> Result<BoolVar, CircuitError> {
+        let computed_root_var = {
+            let proof_var = &proof_var;
+
+            // elem label = H(0, uid, elem)
+            let mut cur_label =
+                Self::DigestGadget::digest_leaf(self, elem_idx_var, proof_var.elem_var)?;
+            for cur_node in proof_var.node_vars.iter() {
+                let input_labels = constrain_sibling_order(
+                    self,
+                    cur_label,
+                    cur_node.sibling1,
+                    cur_node.sibling2,
+                    cur_node.is_left_child,
+                    cur_node.is_right_child,
+                )?;
+                // check that the left child's label is non-zero
+                self.non_zero_gate(input_labels[0])?;
+                cur_label = Self::DigestGadget::digest(self, &input_labels)?;
+            }
+            Ok(cur_label)
+        }?;
+        self.is_equal(root_var, computed_root_var)
+    }
+
+    fn enforce_membership_proof(
+        &mut self,
+        elem_idx_var: Variable,
+        proof_var: Merkle3AryMembershipProofVar,
+        expected_root_var: Variable,
+    ) -> Result<(), CircuitError> {
+        let bool_val =
+            MerkleTreeGadget::<T>::is_member(self, elem_idx_var, proof_var, expected_root_var)?;
+        self.enforce_true(bool_val.into())
     }
 }
